requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
google-generativeai==0.3.2
google-api-python-client==2.111.0
pyyaml==6.0.1
pydantic==2.5.0
httpx==0.25.2
